Migrate aws sdk go v1 to v2 (#3844)
* Migrate aws sdk go v1 to v2

* Migrate aws sdk go v1 to v2

* Adding pending tasks

* Fix the endpoint ports to avoid reconciles

* Addressing Yang's comments

* Resolve merge conflicts

+++ pkg/model/ec2/security_group_ingress.go
@@ -27,9 +27,9 @@ type PrefixList struct {
 type IPPermission struct {
 	IPProtocol string `json:"ipProtocol"`
 	// +optional
-	FromPort *int64 `json:"fromPort,omitempty"`
+	FromPort *int32 `json:"fromPort,omitempty"`
 	// +optional
-	ToPort *int64 `json:"toPort,omitempty"`
+	ToPort *int32 `json:"toPort,omitempty"`
 	// +optional
 	IPRanges []IPRange `json:"ipRanges,omitempty"`
 	// +optional
